Empleado
- Se implementa buscar, inactivar, activar, actualizar y todos

+++ src/main/java/co/edu/usbcali/gestionrh/model/dto/EmpleadoDTO.java
@@ -18,7 +18,7 @@ import lombok.Setter;
 @Setter
 public class EmpleadoDTO implements Serializable {
   private Long id;
-  private Integer idProfesion;
+  private Long idProfesion;
   private String nombres;
   private String apellidos;
   private String correo;
